Update script.js
In this update, I rewrote the whole JS file and in this updated JS file I added the functions where the number of stars changes depending on what number of button the user clicked. I also added a function where the user can't proceed to the thank you state without having to choose how many stars they want to rate. In the instance the user clicks on the submit button without choosing how many stars they want to rate, there will be an error message popping below the rating state.